# Conflicts:
#   backend/api/__init__.py

+++ backend/api/orders/resources.py
@@ -1,18 +1,21 @@
 from flask.views import MethodView
-from . import bp
-from backend.models.orders import Order, OrderSheet
-from backend.extensions import roles_required, unnest
+from flask_login import current_user
+from backend.models import Order
+from backend.models import OrderSheet
+from backend.extensions import Blueprint, roles_required, unnest
 from .schemas import OrderSchema, OrderTableSchema, TimeLineSchema
-from backend.app import db
+from backend.plugins import db
 from flask_smorest import abort
 
-# TODO: Write a method to do sheet ID or latest in a query class
+bp = Blueprint('orders',
+               'orders',
+               description='Request and change orders')
 
 
 @bp.route('/sheet/<sheet_id_or_latest>')
 class Orders(MethodView):
 
-    @roles_required('planner', 'administrator')
+    @roles_required('view-only', 'planner', 'administrator')
     @bp.response(OrderTableSchema)
     @bp.alt_response('UNAUTHORIZED', code=401)
     @bp.alt_response('NOT_FOUND', code=404)
@@ -21,25 +24,13 @@ class Orders(MethodView):
         Get a list of orders from an order sheet.
 
         In case `sheet_id_or_latest` is `latest`, the most recently uploaded
-        order sheet will be used.
+        order sheet will be requested.
 
-        Required roles: planner, administrator
+        Required roles: view-only, planner, administrator
         """
-        try:
-            # Try to parse the sheet_id to an int and get the order sheet
-            order_sheet = OrderSheet.query\
-                .get_or_404(int(sheet_id_or_latest),
-                            description="Order sheet not found")
-        except ValueError:
-            # The sheet_id is not an integer, so check if it is `latest`
-            if sheet_id_or_latest == 'latest':
-                # Get the most recently uploaded sheet
-                order_sheet = OrderSheet.query\
-                    .order_by(OrderSheet.upload_date.desc())\
-                    .first_or_404()
-            else:
-                # Can't understand which sheet is requested
-                return abort(404, message='Order sheet not found')
+        is_view_only = current_user.role == 'view-only'
+        order_sheet = OrderSheet.query.get_sheet_or_404(sheet_id_or_latest,
+                                                        is_view_only)
         return order_sheet
 
     @roles_required('planner', 'administrator')
@@ -59,21 +50,13 @@ class Orders(MethodView):
 
         Required roles: planner, administrator
         """
-        try:
-            # Try to parse the sheet_id to an int and get the order sheet
-            order_sheet = OrderSheet.query \
-                .get_or_404(int(sheet_id_or_latest),
-                            description="Order sheet not found")
-        except ValueError:
-            # The sheet_id is not an integer, so check if it is `latest`
-            if sheet_id_or_latest == 'latest':
-                # Get the most recently uploaded sheet
-                order_sheet = OrderSheet.query\
-                    .order_by(OrderSheet.upload_date.desc())\
-                    .first_or_404()
-            else:
-                # Can't understand which sheet is requested
-                return abort(404, 'Order sheet not found')
+        order_sheet = OrderSheet.query.get_sheet_or_404(sheet_id_or_latest)
+
+        # Published plannings cannot be changed
+        if order_sheet.planning is not None:
+            return abort(400,
+                         message='Order sheet has already been used in a '
+                                 'planning, cannot add a new order to it.')
 
         # Check that when truck_id is set, departure_time also is set
         if ('truck_id' in order) ^ ('departure_time' in order):
@@ -135,6 +118,7 @@ class OrderByID(MethodView):
 
     @roles_required('planner', 'administrator')
     @bp.response(code=204)
+    @bp.alt_response('BAD_REQUEST', code=400)
     @bp.alt_response('UNAUTHORIZED', code=401)
     @bp.alt_response('NOT_FOUND', code=404)
     @bp.alt_response('SERVICE_UNAVAILABLE', code=503)
@@ -148,9 +132,14 @@ class OrderByID(MethodView):
         """
 
         # try to get the order from the orders table
-        order = Order.query.get_or_404(
-            order_id,
-            description='Order not found')
+        order = Order.query.get_or_404(order_id, description='Order not found')
+
+        # Published plannings cannot be changed
+        if order.order_sheet.planning is not None:
+            return abort(400,
+                         message='Order sheet has already been used in a '
+                                 'planning, cannot delete an order from it.')
+
         # delete the order
         db.session.delete(order)
         db.session.commit()
@@ -174,9 +163,15 @@ class OrderByID(MethodView):
         """
         try:
             # Try to parse the sheet_id to an int and get the order
-            order = Order.query.get_or_404(
-                order_id,
-                description='Order not found')
+            order = Order.query.get_or_404(order_id,
+                                           description='Order not found')
+
+            # Published plannings cannot be changed
+            if order.order_sheet.planning is not None:
+                return abort(400,
+                             message='Order sheet has already been used in a '
+                                     'planning, cannot change an order '
+                                     'from it.')
 
             # Make sure the primary key is not changed
             if 'order_number' in req:
@@ -243,27 +238,25 @@ class OrderByID(MethodView):
 @bp.route('/timeline/<sheet_id_or_latest>')
 class DataVisualisation(MethodView):
 
-    @roles_required('planner', 'administrator')
+    @roles_required('view-only', 'planner', 'administrator')
     @bp.response(TimeLineSchema(many=True))
     @bp.alt_response('UNAUTHORIZED', code=401)
     @bp.alt_response('NOT_FOUND', code=404)
     def get(self, sheet_id_or_latest):
-        try:
-            # Try to parse the sheet_id to an int and get the order sheet
-            order_sheet = OrderSheet.query \
-                .get_or_404(int(sheet_id_or_latest),
-                            description="Order sheet not found")
-        except ValueError:
-            # The sheet_id is not an integer, so check if it is `latest`
-            if sheet_id_or_latest == 'latest':
-                # Get the most recently uploaded sheet
-                order_sheet = OrderSheet.query \
-                    .order_by(OrderSheet.upload_date.desc()) \
-                    .first_or_404()
-            else:
-                # Can't understand which sheet is requested
-                return abort(404, message='Order sheet not found')
+        """
+        Gets a the parameters needed for making a timeline.
 
+        In case `sheet_id_or_latest` is `latest`, the most recently uploaded
+        order sheet will be used to create the timeline.
+
+        Required roles: view-only, planner, administrator
+        """
+        # Get the requested order sheet
+        is_view_only = current_user.role == 'view-only'
+        order_sheet = OrderSheet.query.get_sheet_or_404(sheet_id_or_latest,
+                                                        is_view_only)
+
+        # Get all orders from the order sheet that have trucks assigned
         orders = Order.query \
             .filter(Order.sheet_id == order_sheet.id) \
             .filter(Order.truck_id.isnot(None)) \
